refactor: Error/InGame constant 분리
