+++ src/main/java/racingcar/car/constant/Constant.java
@@ -1,0 +1,15 @@
+package racingcar.car.constant;
+
+
+public class Constant {
+    public final static String CAR_IS_NOTMATCHING = "";
+    public final static String CAR_NAMES_DELIMITER = ",";
+    public final static int RANGE_FOR_CAR_MOVING_MIN = 0;
+    public final static int RANGE_FOR_CAR_MOVING_MAX = 9;
+    public final static String GUIDE_FOR_INPUT_CARNAME = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+    public final static String GUIDE_FOR_INPUT_RACINGNUMBER = "시도할 횟수는 몇회인가요?";
+    public final static String PROGRESS_BAR = "-";
+    public final static String PROGRESS_DELIMITER = " : ";
+    public final static String WINNER_DELIMITER = ", ";
+    public final static String WINNER_GUIDE = "최종 우승자: ";
+}
