+++ src/main/java/racingcar/car/model/validator/ValidateRacingNumberInputRangeOverOne.java
@@ -1,7 +1,7 @@
 package racingcar.car.model.validator;
 
-import racingcar.car.Constant;
 import racingcar.car.GameStatus;
+import racingcar.car.constant.ErrorMessageConstant;
 
 public class ValidateRacingNumberInputRangeOverOne implements Validator {
     private final int INPUT_LIMIT = 1;
@@ -16,7 +16,7 @@ public class ValidateRacingNumberInputRangeOverOne implements Validator {
         boolean validateResult = true;
         int inputParseInt = Integer.parseInt(input);
         if (inputParseInt < INPUT_LIMIT) {
-            System.out.println(Constant.ERROR_COMMENT_RACING_NUMBER_RANGE);
+            System.out.println(ErrorMessageConstant.ERROR_COMMENT_RACING_NUMBER_RANGE);
             validateResult = false;
         }
         return validateResult;
